Passing in the function to load facebook to isolate it for future testing, and fix failing test issues
